% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the NOAA dataset}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{the data.frame to be cleaned}
}
\value{
data.frame containing the cleaned data (lubridate ymd format
for the date, latitude and longitude converted to numeric)
}
\description{
This function cleans and readies the NOAA dataset.
}
\examples{
\dontrun{eq_clean_data(rawdataframe)}

}
