% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module3.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Leaflet Earth Quake Map with Customized Label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The earthquake clean data frame.}
}
\value{
A character vector using HTML tags to be used as label for
labelcolumn variable.
}
\description{
eq_create_label creates a more informative popup using HTML tags
to be used as labels in eq_map.
}
\details{
This function put together a character string for each earthquake to be
used as a more informative label. The label contains the following items:
\itemize{
\item Location
\item Magnitude
\item Total Deaths
}
}
\examples{
\dontrun{
load_raw_data() \%>\%
eq_clean_data() \%>\%
eq_location_clean() \%>\%
 dplyr::filter(Country \%in\% c("HONDURAS", "MEXICO") & lubridate::year(Date) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%#'
 eq_map( annot_col = "popup_text")
}

}
