% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_load_data}
\alias{eq_load_data}
\title{Load Pre-processed Earthquakes Data}
\usage{
eq_load_data()
}
\value{
A \code{datframe} of NOAA Significant Earthquakes Data, cleaned with
  the functions \link{eq_clean_data} and \link{eq_location_clean}.
}
\description{
\code{eq_load_data} is to load pre-processed NOAA dataset.
\code{\link{earthquake}} data supplied with this package.
}
\details{
This function is to load the \code{\link{earthquake}} data
supplied with this package.
}
\examples{
library(dplyr)
data <- eq_load_data()

}

