% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module3.R
\name{eq_map}
\alias{eq_map}
\title{Leaflet Map Tools}
\usage{
eq_map(data, annot_col = "Date")
}
\arguments{
\item{data}{The earthquake clean data frame}

\item{annot_col}{The name of the column to be used for annotation}
}
\value{
Returns a leaflet map
}
\description{
eq_map creates a leaflet map
showing the location of the earthquake.
}
\details{
The function maps the epicenters Longitude and
Latitude and annotates each point in a popup window.
}
\examples{
\dontrun{
load_raw_data() \%>\%
eq_clean_data() \%>\%
eq_location_clean() \%>\%
dplyr::filter(Country == "MEXICO" & lubridate::year(Date) >= 2000)
eq_map(data = NOAA_data, annot_col = "Date")
}

}
