% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom for for plotting a time line of earthquakes
with a point for each earthquake}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetics mappings created by aes}

\item{data}{The data to be displayed in this layer}

\item{stat}{not working (no stat functionality)}

\item{position}{Position adjustment}

\item{show.legend}{logical. Should this layer be included in the legends? NA,
the default, includes if any aesthetics are mapped. FALSE never includes,
and TRUE always includes. It can also be a named logical vector to finely
select the aesthetics to display.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.
If TRUE, missing values are silently removed.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{...}{Other arguments to layer}
}
\description{
This function creates a new geom to plot a timeline
for a data range and points for each earthquake
}
\examples{
\dontrun{
dataset \%>\% filter((Year> 2010) &  (Country =="INDIA")) \%>\%
   ggplot(ggplot2::aes(x = Date,
            y = Country,
            color = Deaths,
            size = Mag)) + geom_timeline(alpha = 0.2)
}

}
