% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Label the earthquakes on the timeline}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetics mappings created by ggplot2 aes}

\item{data}{The data to be displayed in this layer}

\item{stat}{not working (no stat functionality)}

\item{position}{Position adjustment, either as a string, or the result of a
call to position adjustment function}

\item{show.legend}{logical. Should this layer be included in the legends? NA,
the default, includes if any aesthetics are mapped. FALSE never includes,
and TRUE always includes. It can also be a named logical vector to finely
select the aesthetics to display}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.
If TRUE, missing values are silently removed}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them}

\item{...}{Other arguments to layer}
}
\description{
geom_timeline_label  labels the n biggest earthquakes
(based on the Richter scale)
}
\examples{
\dontrun{
dataset \%>\% filter(Country \%in\% c("MEXICO", "IRAN") &
     Date \%within\% lubridate::interval(ymd(20000103), ymd(20180104))) \%>\%
     mutate(Country = factor(Country, levels = unique(Country))) \%>\%
     ggplot() +
     geom_timeline(ggplot2::aes(x = Date, y = Country, size = Mag, colour = Deaths)) +
     geom_timeline_label(aes(x = Date, y = Country, label = Location,
                         magnitude = Mag, colour = Deaths, n_max = 7), alpha = 0.5) +
     scale_colour_continuous(name = "Nb. of deaths") +
     theme(legend.position = "bottom") +
     ylab("")
}

}
