% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2.R
\docType{data}
\name{GeomTimeLine}
\alias{GeomTimeLine}
\title{Class GeomTimeline}
\format{
An object of class \code{GeomTimeLine} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimeLine
}
\value{
no return value
}
\description{
Construct a new class 'GeomTimeline' which is a 'ggproto' object
for plotting a time line of earthquakes.
}
\keyword{datasets}
