% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2.R
\docType{data}
\name{GeomTimeLineLabel}
\alias{GeomTimeLineLabel}
\title{GeomTimelineLabel}
\format{
An object of class \code{GeomTimeLineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimeLineLabel
}
\description{
This is the setup required for creating a new geom class.
This GeomTimeLineLabel inherits from a top level class called Geom.
It creates a geom that draws the timeline for a specified date interval
and put points on it for each earthquake. It then creates labels in
order to annotate the biggest earthquakes on the timeline.
}
\keyword{datasets}
