% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load the raw dataset from a text file}
\usage{
load_raw_data(filename = "data/NOAA.txt")
}
\arguments{
\item{filename}{(optional) the raw NOAA data to be cleaned}
}
\value{
data.frame containing the NOAA data
}
\description{
This function reads the U.S. National Oceanographic
and Atmospheric Administration (NOAA) dataset on earthquakes
}
\examples{
\dontrun{load_raw_dataset()}

}
