% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAAdata}
\alias{NOAAdata}
\title{NOAA dataset}
\format{
A data frame with 6218 observations of 39 variables.
}
\source{
\url{https://www.ngdc.noaa.gov/}
}
\usage{
NOAAdata
}
\description{
NOAA dataset
}
\keyword{datasets}
