% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Earthquake data 
\code{eq_clean_data} takes a data frame of NOAA earthquake data and cleans it
up for further analysis.
 This function takes a data frame of NOAA earthquake data, which may be
obtained from NOAA's
\href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{
EarthQuake Database}, or loaded into the envirnoment with \code{data(earthquake)}
and cleans it up for further analysis.  In
particular, it creates a \code{DATE} feature by combing the \code{YEAR},
\code{MONTH}, and \code{DAY} features, and makes sure the \code{LATITUDE} and
\code{LONGITUDE} features are of type \code{numeric}.}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{A data frame of NOAA significant earthquake data, similar to what
can be loaded with \code{data(earthquake)}.}
}
\value{
dataframe. A cleaned version of the earthquake data with a new column \code{DATE}.
}
\description{
Clean Earthquake data 
\code{eq_clean_data} takes a data frame of NOAA earthquake data and cleans it
up for further analysis.
 This function takes a data frame of NOAA earthquake data, which may be
obtained from NOAA's
\href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{
EarthQuake Database}, or loaded into the envirnoment with \code{data(earthquake)}
and cleans it up for further analysis.  In
particular, it creates a \code{DATE} feature by combing the \code{YEAR},
\code{MONTH}, and \code{DAY} features, and makes sure the \code{LATITUDE} and
\code{LONGITUDE} features are of type \code{numeric}.
}

