% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Popup Label for Earthquake Map}
\usage{
eq_create_label(x)
}
\arguments{
\item{x}{The \code{\link{earthquake}} data frame.}
}
\value{
A vector with the HTML-formatted labels. You should include this
vector with the data frame that is sent to \code{\link{eq_map}}.
}
\description{
\code{eq_create_label} creates a more descriptive and HTML-formatted popup
label to be used in \code{\link{eq_map}}.
}
\details{
This function creates a vector of HTML-formatted labels using supplied
\code{\link{earthquake}} data. Variables:
\itemize{
  \item DATE
  \item LOCATION_NAME (as cleaned in the \code{\link{eq_location_clean}}
  function).
  \item EQ_PRIMARY (earthquake magnitude)
  \item TOTAL_DEATHS
}
}
\examples{
library(dplyr); library(lubridate)
earthquakes <- eq_load_data()  
earthquakes \%>\%
 dplyr::filter(COUNTRY == 'INDIA') \%>\%
 dplyr::filter(lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = 'popup_text')

}

