% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Earthquake Location Field
\code{eq_location_clean} takes a data frame of NOAA earthquake data and cleans
the \code{LOCATION_NAME} field.
This function cleans the \code{LOCATION_NAME} observation in a NOAA
Significant Earthquakes data set.  It removes the country name (of form
\code{COUNTRY:}) from the \code{LOCATION_NAME} data (unless the country is
the only name present in that field), and converts the remainder of the field
to Title Case. This will make the location name easier to read when plotting
and mapping.}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{A data frame of NOAA significant earthquake data, similar to what
can be loaded with \code{data(earthquake)}.}
}
\value{
dataframe A cleaned version of the earthquake data with cleaned version of \code{LOCATION_NAME}
column (removing country name, seperated with location name with ":")
}
\description{
Clean Earthquake Location Field
\code{eq_location_clean} takes a data frame of NOAA earthquake data and cleans
the \code{LOCATION_NAME} field.
This function cleans the \code{LOCATION_NAME} observation in a NOAA
Significant Earthquakes data set.  It removes the country name (of form
\code{COUNTRY:}) from the \code{LOCATION_NAME} data (unless the country is
the only name present in that field), and converts the remainder of the field
to Title Case. This will make the location name easier to read when plotting
and mapping.
}

