% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Interactive Earthquake Map}
\usage{
eq_map(x, annot_col = "DATE")
}
\arguments{
\item{x}{The \code{\link{earthquake}} data frame.}

\item{annot_col}{The column to use for the popup annotation.}
}
\value{
Returns an interactive \code{leaflet} map.
}
\description{
\code{eq_map} creates an interactive \code{\link[leaflet]{leaflet}} map
showing the location of earthquakes in the given \code{earthquake} data set.
}
\details{
This function shows an interactive map of the location of the earthquakes in
the given \code{\link{earthquake}} data.  The size of the circles are
proportional to the magnitude of the earthquakes (in the \code{EQ_PRIMARY})
variable. The map is interactive, and when you click on a link, the popup
shows the annotation as specified by the \code{annot_col} variable.
}
\examples{
library(lubridate); library(dplyr)

## use 'DATE' as annotation
earthquakes <- eq_load_data()
earthquakes \%>\%
 dplyr::filter(COUNTRY == 'JAPAN') \%>\%
 dplyr::filter(lubridate::year(DATE) >= 2000)

eq_map(earthquakes, annot_col = 'DATE')

}
\seealso{
\code{\link{eq_create_label}} to create a more useful popup
  annotation in the \code{earthquake} data frame.
}

