% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Earthquake Timeline
\code{geom_timeline} shows a timeline of NOAA Significant earthquakes.
The size of the points is relatative to the earthquakes' magnitude, and
the color is related to the total number of deaths.}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{data mapping}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If TRUE NA are ignored}

\item{show.legend}{ogical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{other parameters}
}
\description{
Earthquake Timeline
\code{geom_timeline} shows a timeline of NOAA Significant earthquakes.
The size of the points is relatative to the earthquakes' magnitude, and
the color is related to the total number of deaths.
}
\section{Aesthetics}{

\code{geom_timeline} undertands the following aesthetics (required
aesthetics are in bold):
\itemize{
 \item x: recommend \code{DATE}
 \item y: recommend \code{COUNTRY}
 \item size: recommend \code{EQ_PRIMARY}
 \item color: recommend \code{TOTAL_DEATHS}
 \item alpha
 \item shape
}
}
\examples{
library(dplyr); library(ggplot2)
quakes <- eq_load_data()

quakes \%>\%
  dplyr::filter(COUNTRY \%in\% c('USA', 'INDIA')) \%>\%
  dplyr::filter(DATE > '2005-05-01') \%>\%
  ggplot() +
  geom_timeline(aes(x = DATE, y = COUNTRY, color = TOTAL_DEATHS,
                    size = EQ_PRIMARY)) +
  scale_size_continuous(name = 'Richter scale value') +
  scale_color_continuous(name = 'Deaths') 
  
}

