% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Label Earthquakes on Timeline}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{data mapping}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If TRUE NA are ignored}

\item{show.legend}{ogical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{other parameters}
}
\description{
\code{geom_timeline_label} works best when used with
\code{\link{geom_timeline}}, labeling the top \code{n} earthquakes, by
magnitude, with a specified label field.  By default, the labels are for
the top 5 earthquakes for each country specified, however, the user may
adjust this with the \code{n_max} aesthetic.
}
\section{Aesthetics}{

\code{geom_timeline_label} undertands the following aesthetics (required
aesthetics are in bold):
\itemize{
 \item \strong{x}: recommend \code{DATE}
 \item \strong{label}: recommend \code{LOCATION_NAME}
 \item \strong{magnitude}: recommend \code{EQ_PRIMARY}
 \item y: recommend \code{COUNTRY}
 \item n_max: default 5. Top \code{n} earthquakes to label,
       sorted by magnitude.
 \item color
 \item linetype
 \item size
 \item alpha
}
}
\examples{
library(dplyr); library(ggplot2)
earthquakes <- eq_load_data()

earthquakes \%>\%
  dplyr::filter(COUNTRY \%in\% c('USA', 'INDIA')) \%>\%
  dplyr::filter(DATE > '2005-05-01') \%>\%
  ggplot() +
  geom_timeline(aes(x = DATE, y = COUNTRY, color = TOTAL_DEATHS,
                    size = EQ_PRIMARY)) +
  geom_timeline_label(aes(x = DATE, y = COUNTRY, magnitude = EQ_PRIMARY,
                         label = LOCATION_NAME, n_max = 5)) +
  scale_size_continuous(name = 'Richter scale value') +
  scale_color_continuous(name = 'Deaths') +
  theme_eq()
  
}

